+++ mapbox_maps_flutter_mobile/lib/src/mapbox_maps_flutter_mobile.dart
@@ -18,8 +18,7 @@ base class MapboxMapsFlutterMobile extends MapboxMapsFlutterPlatform
     void Function(MapEvent)? onMapEvent,
     MapOptions? mapOptions,
     bool? textureView,
-    AndroidPlatformViewHostingMode androidHostingMode =
-        AndroidPlatformViewHostingMode.VD,
+    required AndroidPlatformViewHostingMode androidHostingMode,
     Set<Factory<OneSequenceGestureRecognizer>>? gestureRecognizers,
     bool? isOpaque = true,
   }) {